SunOS: silences compiler warnings

+++ src/detection/memory/memory_sunos.c
@@ -3,8 +3,8 @@
 
 const char* ffDetectMemory(FFMemoryResult* ram)
 {
-    ram->bytesTotal = sysconf(_SC_PHYS_PAGES) * instance.state.platform.sysinfo.pageSize;
-    ram->bytesUsed = ram->bytesTotal - sysconf(_SC_AVPHYS_PAGES) * instance.state.platform.sysinfo.pageSize;
+    ram->bytesTotal = (uint64_t) sysconf(_SC_PHYS_PAGES) * instance.state.platform.sysinfo.pageSize;
+    ram->bytesUsed = ram->bytesTotal - (uint64_t) sysconf(_SC_AVPHYS_PAGES) * instance.state.platform.sysinfo.pageSize;
 
     return NULL;
 }
